Moved code to relevant files
Got rid of redundant codes by breaking down into mehods
