+++ src/index.js
@@ -2,13 +2,10 @@ import './css/main.css';
 import './css/modal.css';
 import initHomePage from './home';
 import initModal from './modal';
-import Todo from './models/todo';
-import Project from './models/project';
+import { Todo, AddTodoHTML } from './models/todo';
+import { Project, NewProjectHTML, allProjects } from './models/project';
 
-// Refacture later
-const allProjects = [];
-
-// Home Page New Todo Button Brings up Modal
+// Home Page  + New Todo Button Brings up Modal
 const NewTodo = (e) => {
   const todoBtn = e.target;
   const todoBtnID = parseInt(todoBtn.dataset.id, Math.radix);
@@ -19,31 +16,12 @@ const NewTodo = (e) => {
   initModal.show(e);
 };
 
-// Function for Modal New Project Button
+// Function for Modal Button New Project Button
 function AddProject() {
   const id = allProjects.length + 1;
   const title = initModal.getProjectTitleInput();
 
-  const projectContent = document.createElement('div');
-  projectContent.setAttribute('id', id);
-  projectContent.classList.add('container');
-
-  const projectTodoContent = document.createElement('div');
-  
-  const header = document.createElement('h3');
-  header.innerHTML = title;
-
-  const newTodoBtn = document.createElement('button');
-  newTodoBtn.classList.add('button-todo');
-  newTodoBtn.classList.add('btn-plus');
-  newTodoBtn.setAttribute('data-id', id.toString());
-  newTodoBtn.innerHTML = 'New Todo';
-
-
-  projectContent.appendChild(header);
-  projectContent.appendChild(projectTodoContent);
-  projectContent.appendChild(newTodoBtn);
-  document.body.appendChild(projectContent);
+  const projectTodoContent = NewProjectHTML(id, title);
 
   const project = Project(id, title);
   project.setProjElement(projectTodoContent);
@@ -53,7 +31,7 @@ function AddProject() {
   initModal.hide();
 }
 
-//Function for Modal New Todo Button
+//Function for Modal Button New Todo Button
 function AddTodo() {
   const project = initHomePage.getSelectedProject();
   const projectDiv = project.getProjElement();
@@ -63,31 +41,17 @@ function AddTodo() {
   const priority = initModal.getTodoPriorityInput();
   
   const todo = Todo(title, desc, dueDate, priority);
-  project.addTodo(todo);
+  project.addTodo(todo, projectDiv);
   
-  const todoContent = document.createElement('div');
-  todoContent.setAttribute('id', todo.id);
-  todoContent.classList.add('todo');
-  const todoChecked = document.createElement('input');
-  todoChecked.classList.add('inline-block');
-  todoChecked.setAttribute('type', 'checkbox');
-  const todoTitle = document.createElement('div');
-  todoTitle.innerHTML = todo.title;
-  const todoDueDate = document.createElement('div');
-  todoDueDate.innerHTML = todo.dueDate;
-
-  todoContent.appendChild(todoChecked);
-  todoContent.appendChild(todoTitle);
-  todoContent.appendChild(todoDueDate);
-  projectDiv.appendChild(todoContent);
+  AddTodoHTML(todo, projectDiv);
 
   initModal.hide();
 }
 
-document.body.appendChild(initHomePage.init());
 document.body.appendChild(initModal.init());
+document.body.appendChild(initHomePage.init());
 
-// Add eventlistner on click NewProject to New Project HTML Button
+// Add eventlistners 'on click' To all buttons... the two Modal buttons and the homepage New Project Button
 initHomePage.newProject(initModal.show);
 initModal.addProjectBtn(AddProject);
 initModal.addTodoBtn(AddTodo);
